requirements.txt
packaging==24.2
wheel @ file:///opt/homebrew/Cellar/python%403.13/3.13.2/libexec/wheel-0.45.1-py3-none-any.whl#sha256=b9235939e2096903717cb6bfc132267f8a7e46deb2ec3ef9c5e234ea301795d0
rich
langchain
langchain_experimental
python-dotenv
langchain_openai
pandas
numpy
tabulate
